frontend/fixAddReviewerModal

+++ frontend/src/components/AddReviewerModal.svelte
@@ -35,10 +35,12 @@
     let requestPath;
     let requests = null;
     let pending_reviewer = [];
+    let new_reviewers = [];
 
 
     function addReviewer(u) {
         reviewers = reviewers.concat([u])
+        new_reviewers = new_reviewers.concat([u])
     }
 
     function apply_query(q) {
@@ -66,7 +68,7 @@
         let data = {
             reviewer_id: reviewer.id
         }
-        requestPath = "/api/categories/" + paper.category_id + "/entries/"+ paper.id +"/process/requests"
+        requestPath = "/api/categories/" + paper.category_id + "/entries/" + paper.id + "/process/requests"
         return fetch(requestPath, {
             method: "POST",
             headers: {
@@ -78,16 +80,22 @@
             .then(resp => {
                 if (resp.status < 200 || resp.status >= 300) {
                     error = "" + resp.status + ": " + resp.message;
-                    reviewers = reviewers.filter(r => r !== reviewer)
                     console.log(error);
-                } else {}
+                } else {
+                    new_reviewers = new_reviewers.filter(new_reviewer => new_reviewer !== reviewer)
+                    pending_reviewer = pending_reviewer.concat([reviewer.id])
+                    loadRequests()
+                    hide()
+                }
             })
             .catch(err => console.log(err));
     }
 
     function loadRequests() {
         requests = null;
-        fetch("/api/categories/"+ paper.category_id + "/entries/" + paper.id + "/process/requests")
+        reviewers = [];
+        pending_reviewer = [];
+        fetch("/api/categories/" + paper.category_id + "/entries/" + paper.id + "/process/requests")
             .then(resp => resp.json())
             .then(resp => {
                 if (resp.status < 200 || resp.status >= 300) {
@@ -102,21 +110,19 @@
     }
 
     function filterUsers(r) {
-        reviewers = users.filter( u => u.id === r.reviewer_id)
-        requests.filter(request => {
-            if(request.state === "PENDING") {
-                pending_reviewer = pending_reviewer.concat(request.reviewer_id)
-            }
-        })
+        reviewers = reviewers.concat(users.filter(u => u.id === r.reviewer_id))
+        if (r.state === "PENDING") {
+            pending_reviewer = pending_reviewer.concat(r.reviewer_id)
+        }
     }
 
     function retractRequest(reviewer) {
-        if(pending_reviewer.includes(reviewer)) {
-            let reviewer_request = requests.filter(request => request.reviewer_id)
-            fetch("/api/categories/" + paper.category_id + "/entries/" + paper.id + "/process/requests/" + reviewer_request[0], {
+        if (pending_reviewer.includes(reviewer.id) && !new_reviewers.includes(reviewer.id)) {
+            let reviewer_request = requests.filter(request => request.reviewer_id === reviewer.id)
+            fetch("/api/categories/" + paper.category_id + "/entries/" + paper.id + "/process/requests/" + reviewer_request[0].id, {
                 method: 'DELETE',
             })
-                .then((response) => console.log(response))
+                .then((response) => response.json())
                 .then((resp) => {
                     if (resp.status < 200 || resp.status >= 300) {
                         error = "" + resp.status + ": " + resp.message;
@@ -126,20 +132,20 @@
                 .catch(err => console.log(err))
 
         } else {
-            reviewers = reviewers.filter(e => e !== reviewer)
+            new_reviewers = new_reviewers.filter(e => e !== reviewer)
         }
+        reviewers = reviewers.filter(e => e !== reviewer)
     }
 
     onMount(() => {
         loadUsers()
         current_user = JSON.parse(Cookies.get("current-user") ?? "{}")
-        console.log(current_user)
     });
 
     function sendRequests() {
-        reviewers.map(reviewer => createDirectRequest(reviewer))
-        hide()
+        new_reviewers.map(reviewer => createDirectRequest(reviewer))
     }
+
 
 </script>
 
@@ -184,36 +190,41 @@
                     {#each reviewers as r }
                         <TableBodyRow>
                             <TableBodyCell>{r.firstName + " " + r.lastName}</TableBodyCell>
-                            <TableBodyCell>
-                                <div class="flex flex-wrap items-center gap-2">
-                                    {#if (pending_reviewer.includes(r))}
+                            {#if (pending_reviewer.includes(r.id) || new_reviewers.includes(r))}
+                                <TableBodyCell>
+                                    PENDING
+                                </TableBodyCell>
+                                <TableBodyCell>
+                                    <div class="flex flex-wrap items-center gap-2">
                                         <Button pill class="!p-2" outline color="red"
                                                 on:click={() => retractRequest(r)}>
                                             <svg class="w-5 h-5" xmlns="http://www.w3.org/2000/svg" x="0px" y="0px"
                                                  width="32px" height="32px" viewBox="0 0 64 64"
                                                  xml:space="preserve">
-                                      <g>
-                                        <line fill="none" stroke="#000000" stroke-width="2" stroke-miterlimit="10" x1="18.947"
-                                              y1="17.153" x2="45.045"
-                                              y2="43.056"/>
-                                      </g>
+                                          <g>
+                                            <line fill="none" stroke="#000000" stroke-width="2" stroke-miterlimit="10"
+                                                  x1="18.947"
+                                                  y1="17.153" x2="45.045"
+                                                  y2="43.056"/>
+                                          </g>
                                                 <g>
-                                        <line fill="none" stroke="#000000" stroke-width="2" stroke-miterlimit="10" x1="19.045"
-                                              y1="43.153" x2="44.947"
-                                              y2="17.056"/>
-                                      </g>
-                                  </svg>
+                                            <line fill="none" stroke="#000000" stroke-width="2" stroke-miterlimit="10"
+                                                  x1="19.045"
+                                                  y1="43.153" x2="44.947"
+                                                  y2="17.056"/>
+                                          </g>
+                                      </svg>
                                         </Button>
-                                    {/if}
-                                </div>
-                            </TableBodyCell>
+                                    </div>
+                                </TableBodyCell>
+                            {/if}
                         </TableBodyRow>
                     {/each}
                 </TableBody>
             </Table>
         </div>
         <Footer class="bottom-0 left-0 z-20 w-full">
-            <Button class="w-full" color="primary" size="sm" type="submit" on:click={() => sendRequests()}>
+            <Button class="w-full" color="primary" size="sm" type="submit" on:click|once={() => sendRequests()}>
                 Send Requests
             </Button>
         </Footer>
